+++ include/GafferSceneUI/TypeIds.h
@@ -55,6 +55,7 @@ enum TypeId
 	CameraToolTypeId = 110661,
 	UVViewTypeId = 110662,
 	UVSceneTypeId = 110663,
+	HistoryPathTypeId = 110664,
 
 	LastTypeId = 110700
 };